rework interface code to remove dedup code
