+++ duatic_dynaarm_controllers/src/dynaarm_status_broadcaster.cpp
@@ -56,21 +56,21 @@ controller_interface::InterfaceConfiguration StatusBroadcaster::state_interface_
     config.names.emplace_back(joint + "/" + hardware_interface::HW_IF_POSITION);
     config.names.emplace_back(joint + "/" + hardware_interface::HW_IF_VELOCITY);
     config.names.emplace_back(joint + "/" + hardware_interface::HW_IF_EFFORT);
-    config.names.emplace_back(joint + "/motor_temperature_system");
-    config.names.emplace_back(joint + "/motor_temperature_coil_A");
-    config.names.emplace_back(joint + "/motor_temperature_coil_B");
-    config.names.emplace_back(joint + "/motor_temperature_coil_C");
-    config.names.emplace_back(joint + "/motor_bus_voltage");
+    config.names.emplace_back(joint + "/temperature_system");
+    config.names.emplace_back(joint + "/temperature_coil_A");
+    config.names.emplace_back(joint + "/temperature_coil_B");
+    config.names.emplace_back(joint + "/temperature_coil_C");
+    config.names.emplace_back(joint + "/bus_voltage");
 
     config.names.emplace_back(joint + "/position_commanded");
     config.names.emplace_back(joint + "/velocity_commanded");
     config.names.emplace_back(joint + "/effort_commanded");
 
-    config.names.emplace_back(joint + "/motor_current_d");
-    config.names.emplace_back(joint + "/motor_current_q");
-    config.names.emplace_back(joint + "/motor_current_coil_A");
-    config.names.emplace_back(joint + "/motor_current_coil_B");
-    config.names.emplace_back(joint + "/motor_current_coil_C");
+    config.names.emplace_back(joint + "/current_d");
+    config.names.emplace_back(joint + "/current_q");
+    config.names.emplace_back(joint + "/current_coil_A");
+    config.names.emplace_back(joint + "/current_coil_B");
+    config.names.emplace_back(joint + "/current_coil_C");
   }
   return config;
 }
@@ -169,56 +169,56 @@ StatusBroadcaster::on_activate([[maybe_unused]] const rclcpp_lifecycle::State& p
     return controller_interface::CallbackReturn::FAILURE;
   }
 
-  if (!controller_interface::get_ordered_interfaces(state_interfaces_, params_.joints, "motor_temperature_system",
+  if (!controller_interface::get_ordered_interfaces(state_interfaces_, params_.joints, "temperature_system",
                                                     joint_temperature_system_interfaces_)) {
-    RCLCPP_WARN(get_node()->get_logger(), "Could not get ordered interfaces - motor_temperature_system");
+    RCLCPP_WARN(get_node()->get_logger(), "Could not get ordered interfaces - Temperature_system");
     return controller_interface::CallbackReturn::FAILURE;
   }
-  if (!controller_interface::get_ordered_interfaces(state_interfaces_, params_.joints, "motor_temperature_coil_A",
+  if (!controller_interface::get_ordered_interfaces(state_interfaces_, params_.joints, "temperature_coil_A",
                                                     joint_temperature_phase_a_interfaces_)) {
-    RCLCPP_WARN(get_node()->get_logger(), "Could not get ordered interfaces - motor_temperature_coil_A");
+    RCLCPP_WARN(get_node()->get_logger(), "Could not get ordered interfaces - temperature_coil_A");
     return controller_interface::CallbackReturn::FAILURE;
   }
-  if (!controller_interface::get_ordered_interfaces(state_interfaces_, params_.joints, "motor_temperature_coil_B",
+  if (!controller_interface::get_ordered_interfaces(state_interfaces_, params_.joints, "temperature_coil_B",
                                                     joint_temperature_phase_b_interfaces_)) {
-    RCLCPP_WARN(get_node()->get_logger(), "Could not get ordered interfaces - motor_temperature_coil_B");
+    RCLCPP_WARN(get_node()->get_logger(), "Could not get ordered interfaces - temperature_coil_B");
     return controller_interface::CallbackReturn::FAILURE;
   }
-  if (!controller_interface::get_ordered_interfaces(state_interfaces_, params_.joints, "motor_temperature_coil_C",
+  if (!controller_interface::get_ordered_interfaces(state_interfaces_, params_.joints, "temperature_coil_C",
                                                     joint_temperature_phase_c_interfaces_)) {
-    RCLCPP_WARN(get_node()->get_logger(), "Could not get ordered interfaces - motor_temperature_coil_C");
+    RCLCPP_WARN(get_node()->get_logger(), "Could not get ordered interfaces - temperature_coil_C");
     return controller_interface::CallbackReturn::FAILURE;
   }
-  if (!controller_interface::get_ordered_interfaces(state_interfaces_, params_.joints, "motor_bus_voltage",
+  if (!controller_interface::get_ordered_interfaces(state_interfaces_, params_.joints, "bus_voltage",
                                                     joint_bus_voltage_interfaces_)) {
-    RCLCPP_WARN(get_node()->get_logger(), "Could not get ordered interfaces - motor_bus_voltage");
+    RCLCPP_WARN(get_node()->get_logger(), "Could not get ordered interfaces - bus_voltage");
     return controller_interface::CallbackReturn::FAILURE;
   }
 
-  if (!controller_interface::get_ordered_interfaces(state_interfaces_, params_.joints, "motor_current_d",
+  if (!controller_interface::get_ordered_interfaces(state_interfaces_, params_.joints, "current_d",
                                                     joint_current_d_interfaces_)) {
-    RCLCPP_WARN(get_node()->get_logger(), "Could not get ordered interfaces - motor_current_d");
+    RCLCPP_WARN(get_node()->get_logger(), "Could not get ordered interfaces - current_d");
     return controller_interface::CallbackReturn::FAILURE;
   }
-  if (!controller_interface::get_ordered_interfaces(state_interfaces_, params_.joints, "motor_current_q",
+  if (!controller_interface::get_ordered_interfaces(state_interfaces_, params_.joints, "current_q",
                                                     joint_current_q_interfaces_)) {
-    RCLCPP_WARN(get_node()->get_logger(), "Could not get ordered interfaces - motor_current_q");
+    RCLCPP_WARN(get_node()->get_logger(), "Could not get ordered interfaces - current_q");
     return controller_interface::CallbackReturn::FAILURE;
   }
 
-  if (!controller_interface::get_ordered_interfaces(state_interfaces_, params_.joints, "motor_current_coil_A",
+  if (!controller_interface::get_ordered_interfaces(state_interfaces_, params_.joints, "current_coil_A",
                                                     joint_current_phase_a_interfaces_)) {
-    RCLCPP_WARN(get_node()->get_logger(), "Could not get ordered interfaces - motor_current_coil_A");
+    RCLCPP_WARN(get_node()->get_logger(), "Could not get ordered interfaces - current_coil_A");
     return controller_interface::CallbackReturn::FAILURE;
   }
-  if (!controller_interface::get_ordered_interfaces(state_interfaces_, params_.joints, "motor_current_coil_B",
+  if (!controller_interface::get_ordered_interfaces(state_interfaces_, params_.joints, "current_coil_B",
                                                     joint_current_phase_b_interfaces_)) {
-    RCLCPP_WARN(get_node()->get_logger(), "Could not get ordered interfaces - motor_current_coil_B");
+    RCLCPP_WARN(get_node()->get_logger(), "Could not get ordered interfaces - current_coil_B");
     return controller_interface::CallbackReturn::FAILURE;
   }
-  if (!controller_interface::get_ordered_interfaces(state_interfaces_, params_.joints, "motor_current_coil_C",
+  if (!controller_interface::get_ordered_interfaces(state_interfaces_, params_.joints, "current_coil_C",
                                                     joint_current_phase_c_interfaces_)) {
-    RCLCPP_WARN(get_node()->get_logger(), "Could not get ordered interfaces - motor_current_coil_C");
+    RCLCPP_WARN(get_node()->get_logger(), "Could not get ordered interfaces - current_coil_C");
     return controller_interface::CallbackReturn::FAILURE;
   }
   return controller_interface::CallbackReturn::SUCCESS;
