--- duatic_dynaarm_driver/duatic_dynaarm_driver/include/duatic_dynaarm_driver/dynaarm_mock_hardware_interface.hpp
@@ -1,97 +1,0 @@
-/*
- * Copyright 2026 Duatic AG
- *
- * Redistribution and use in source and binary forms, with or without modification, are permitted provided that the
- * following conditions are met:
- *
- * 1. Redistributions of source code must retain the above copyright notice, this list of conditions and the following
- * disclaimer.
- *
- * 2. Redistributions in binary form must reproduce the above copyright notice, this list of conditions and the
- * following disclaimer in the documentation and/or other materials provided with the distribution.
- *
- * 3. Neither the name of the copyright holder nor the names of its contributors may be used to endorse or promote
- * products derived from this software without specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES,
- * INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
- * SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY,
- * WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-#pragma once
-
-// System
-#include <limits>
-#include <memory>
-#include <string>
-#include <vector>
-#include <unordered_map>
-
-// ros2_control hardware_interface
-#include <rclcpp/rclcpp.hpp>
-#include "hardware_interface/hardware_info.hpp"
-#include "hardware_interface/system_interface.hpp"
-#include "hardware_interface/types/hardware_interface_return_values.hpp"
-
-// ROS
-#include <rclcpp/macros.hpp>
-#include <rclcpp_lifecycle/node_interfaces/lifecycle_node_interface.hpp>
-#include <rclcpp_lifecycle/state.hpp>
-#include <ament_index_cpp/get_package_share_directory.hpp>
-
-// sdk
-#include <duatic_ros2control_hardware/duadrive_interface_mock.hpp>
-#include <duatic_ros2control_hardware/interface_utils.hpp>
-#include <duatic_dynaarm_driver/types.hpp>
-
-namespace duatic_dynaarm_driver
-{
-class DynaArmMockHardwareInterface : public hardware_interface::SystemInterface
-{
-public:
-  RCLCPP_SHARED_PTR_DEFINITIONS(DynaArmMockHardwareInterface)
-  DynaArmMockHardwareInterface();
-  ~DynaArmMockHardwareInterface();
-
-  std::vector<hardware_interface::InterfaceDescription> export_unlisted_state_interface_descriptions() override;
-  std::vector<hardware_interface::InterfaceDescription> export_unlisted_command_interface_descriptions() override;
-
-  hardware_interface::CallbackReturn
-  on_init(const hardware_interface::HardwareComponentInterfaceParams& system_info) override;
-  hardware_interface::CallbackReturn on_configure(const rclcpp_lifecycle::State& previous_state) override;
-  hardware_interface::CallbackReturn on_activate(const rclcpp_lifecycle::State& previous_state) override;
-  hardware_interface::CallbackReturn on_deactivate(const rclcpp_lifecycle::State& previous_state) override;
-
-  hardware_interface::return_type read(const rclcpp::Time& time, const rclcpp::Duration& period) override;
-  hardware_interface::return_type write(const rclcpp::Time& time, const rclcpp::Duration& period) override;
-
-  hardware_interface::return_type prepare_command_mode_switch(const std::vector<std::string>& start_interfaces,
-                                                              const std::vector<std::string>& stop_interfaces) override;
-
-  hardware_interface::return_type perform_command_mode_switch(const std::vector<std::string>& start_interfaces,
-                                                              const std::vector<std::string>& stop_interfaces) override;
-
-private:
-  std::vector<duatic_ros2control_hardware::DuaDriveInterfaceMock::UniquePtr> drives_;
-  std::vector<CoupledJointState> state_coupled_kinematics_;
-  std::vector<SerialJointState> state_serial_kinematics_;
-
-  std::vector<CoupledCommand> commands_coupled_kinematics_;
-  std::vector<SerialCommand> commands_serial_kinematics_;
-
-  std::unordered_map<std::string, duatic_ros2control_hardware::SupportedVariant> state_interface_pre_mapping_;
-  std::unordered_map<std::string, duatic_ros2control_hardware::SupportedVariant> command_interface_pre_mapping_;
-
-  duatic_ros2control_hardware::StateInterfaceMapping state_interface_mapping_;
-  duatic_ros2control_hardware::CommandInterfaceMapping command_interface_mapping_;
-
-  hardware_interface::CommandInterface::SharedPtr freeze_mode_interface_;
-
-  rclcpp::Logger logger_;
-};
-
-}  // namespace duatic_dynaarm_driver
