+++ duatic_dynaarm_driver/duatic_dynaarm_driver/src/dynaarm_hardware_interface.cpp
@@ -22,241 +22,12 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include "duatic_dynaarm_driver/dynaarm_hardware_interface.hpp"
-#include "duatic_dynaarm_driver/kinematic_translation.hpp"
-
-using namespace duatic_ros2control_hardware;  // NOLINT(build/namespaces)
+#include "duatic_dynaarm_driver/dynaarm_hardware_interface_base.hpp"
+#include <duatic_ros2control_hardware/duadrive_interface.hpp>
 
 namespace duatic_dynaarm_driver
 {
-
-DynaArmHardwareInterface::DynaArmHardwareInterface() : logger_(rclcpp::get_logger("DynaArmHardwareInterface"))
-{
-}
-
-std::vector<hardware_interface::InterfaceDescription>
-DynaArmHardwareInterface::export_unlisted_state_interface_descriptions()
-{
-  // This is rather simple - we just append all state interface descriptions of each drive and are done
-  std::vector<hardware_interface::InterfaceDescription> state_interfaces;
-
-  for (const auto& drive : drives_) {
-    RCLCPP_INFO_STREAM(logger_, "Setting up state interfaces for drive: " << drive->get_name());
-    const auto drive_state_interfaces = drive->get_state_interface_descriptions();
-    state_interfaces.insert(state_interfaces.end(), drive_state_interfaces.begin(), drive_state_interfaces.end());
-  }
-
-  // More interesting is now the internal state data mapping
-  for (std::size_t i = 0; i < drives_.size(); i++) {
-    auto& drive = drives_[i];
-    auto& state = state_serial_kinematics_[i];
-
-    auto state_mapping = drive->get_default_state_mapping();
-    // Now comes the magic - we replace the "position, velocity, acceleration, torque" fields (+ their commanded
-    // counterparts) with our local ones that have the translated kinematics
-    state_mapping[get_interface_name(drive->get_name(), hardware_interface::HW_IF_POSITION)] = &state.position;
-    state_mapping[get_interface_name(drive->get_name(), hardware_interface::HW_IF_VELOCITY)] = &state.velocity;
-    state_mapping[get_interface_name(drive->get_name(), hardware_interface::HW_IF_ACCELERATION)] = &state.acceleration;
-    state_mapping[get_interface_name(drive->get_name(), hardware_interface::HW_IF_EFFORT)] = &state.torque;
-
-    state_mapping[get_interface_name(drive->get_name(), "position_commanded")] = &state.position_commanded;
-    state_mapping[get_interface_name(drive->get_name(), "velocity_commanded")] = &state.velocity_commanded;
-    state_mapping[get_interface_name(drive->get_name(), "acceleration_commanded")] = &state.acceleration_commanded;
-    state_mapping[get_interface_name(drive->get_name(), "effort_commanded")] = &state.torque_commanded;
-  }
-  return state_interfaces;
-}
-std::vector<hardware_interface::InterfaceDescription>
-DynaArmHardwareInterface::export_unlisted_command_interface_descriptions()
-{
-  std::vector<hardware_interface::InterfaceDescription> command_interfaces;
-
-  for (const auto& drive : drives_) {
-    RCLCPP_INFO_STREAM(logger_, "Setting up state interfaces for drive: " << drive->get_name());
-    const auto drive_command_interfaces = drive->get_command_interface_descriptions();
-    command_interfaces.insert(command_interfaces.end(), drive_command_interfaces.begin(),
-                              drive_command_interfaces.end());
-  }
-
-  for (std::size_t i = 0; i < drives_.size(); i++) {
-    auto& drive = drives_[i];
-    auto& cmd = commands_serial_kinematics_[i];
-
-    auto cmd_mapping = drive->get_default_command_mapping();
-    // Now comes the magic - we replace the "position, velocity, acceleration, torque" fields (+ their commanded
-    // counterparts) with our local ones that have the translated kinematics
-    cmd_mapping[get_interface_name(drive->get_name(), hardware_interface::HW_IF_POSITION)] = &cmd.position;
-    cmd_mapping[get_interface_name(drive->get_name(), hardware_interface::HW_IF_VELOCITY)] = &cmd.velocity;
-    cmd_mapping[get_interface_name(drive->get_name(), hardware_interface::HW_IF_ACCELERATION)] = &cmd.acceleration;
-    cmd_mapping[get_interface_name(drive->get_name(), hardware_interface::HW_IF_EFFORT)] = &cmd.torque;
-  }
-
-  return command_interfaces;
-}
-
-hardware_interface::CallbackReturn
-DynaArmHardwareInterface::on_init(const hardware_interface::HardwareComponentInterfaceParams& system_info)
-{
-  const auto arm_name = system_info.hardware_info.name;
-  // The logger is now a child logger with a more descriptive name
-  logger_ = logger_.get_child(arm_name);
-
-  const auto ethercat_bus = system_info.hardware_info.hardware_parameters.at("ethercat_bus");
-  // We obtain information about configured joints and create DuaDriveInterface instances from them
-  // and initialize them
-  for (const auto& joint : system_info.hardware_info.joints) {
-    const auto address = std::stoi(joint.parameters.at("address"));
-    const auto joint_name = joint.name;
-    const std::string device_file_path = joint.parameters.at("drive_parameter_file");
-    const std::string default_parameter_file_path = joint.parameters.at("drive_parameter_folder_default") + "/example_"
-                                                                                                            "drive_"
-                                                                                                            "config."
-                                                                                                            "yaml";
-
-    RCLCPP_INFO_STREAM(logger_, "Setup drive instance for joint: " << joint_name << " on ethercat bus:" << ethercat_bus
-                                                                   << " at address: " << address);
-    drives_.emplace_back(std::make_unique<DuaDriveInterface>(logger_));
-    // As we need to apply the kinematic translation we need some place to store the corresponding data
-    // TODO(firesurfer) we could ellide copies if we would directly use pointers on the state interface
-    state_coupled_kinematics_.emplace_back(CoupledJointState{});
-    state_serial_kinematics_.emplace_back(SerialJointState{});
-
-    commands_coupled_kinematics_.emplace_back(SerialCommand{});
-    commands_serial_kinematics_.emplace_back(CoupledCommand{});
-    // Init doesn't really do anything apart from setting parameters
-    drives_.back()->init(DuaDriveInterfaceParameters{ .ethercat_bus = ethercat_bus,
-                                                      .joint_name = joint_name,
-                                                      .drive_parameter_file_path = device_file_path,
-                                                      .drive_default_parameter_file_path = default_parameter_file_path,
-                                                      .device_address = address });
-  }
-  return hardware_interface::CallbackReturn::SUCCESS;
-}
-
-hardware_interface::CallbackReturn
-DynaArmHardwareInterface::on_configure([[maybe_unused]] const rclcpp_lifecycle::State& previous_state)
-{
-  for (auto& drive : drives_) {
-    // Call configure for each drive and propagate errors if necessary
-    // We currently treat every error that can happen in this stage as fatal
-    if (drive->configure() != hardware_interface::CallbackReturn::SUCCESS) {
-      RCLCPP_FATAL_STREAM(logger_, "Failed to 'configure' drive: " << drive->get_name() << ". Aborting startup!");
-      return hardware_interface::CallbackReturn::FAILURE;
-    }
-  }
-  return hardware_interface::CallbackReturn::SUCCESS;
-}
-hardware_interface::CallbackReturn
-DynaArmHardwareInterface::on_activate([[maybe_unused]] const rclcpp_lifecycle::State& previous_state)
-{
-  for (auto& drive : drives_) {
-    // Call activate for each drive and propagate errors if necessary
-    // We currently treat every error that can happen in this stage as fatal
-    if (drive->activate() != hardware_interface::CallbackReturn::SUCCESS) {
-      RCLCPP_FATAL_STREAM(logger_, "Failed to 'activate' drive: " << drive->get_name() << ". Aborting startup!");
-      return hardware_interface::CallbackReturn::FAILURE;
-    }
-  }
-  return hardware_interface::CallbackReturn::SUCCESS;
-}
-
-hardware_interface::CallbackReturn
-DynaArmHardwareInterface::on_deactivate(const rclcpp_lifecycle::State& /*previous_state*/)
-{
-  for (auto& drive : drives_) {
-    // Call deactivate for each drive and propagate errors if necessary
-    if (drive->deactivate() != hardware_interface::CallbackReturn::SUCCESS) {
-      RCLCPP_FATAL_STREAM(logger_, "Failed to 'deactivate' drive: " << drive->get_name() << ". Aborting startup!");
-      return hardware_interface::CallbackReturn::FAILURE;
-    }
-  }
-  return hardware_interface::CallbackReturn::SUCCESS;
-}
-
-hardware_interface::return_type DynaArmHardwareInterface::read([[maybe_unused]] const rclcpp::Time& time,
-                                                               [[maybe_unused]] const rclcpp::Duration& period)
-{
-  // TODO(firesurfer) - replace with std::views::zip (or own implementation) when available
-  for (std::size_t i = 0; i < drives_.size(); i++) {
-    auto& drive = drives_[i];
-    auto& state = state_coupled_kinematics_[i];
-    // Try to read from each drive - in case of an error
-    if (drive->read() != hardware_interface::return_type::OK) {
-      RCLCPP_ERROR_STREAM(logger_, "Failed to 'read' from drive: " << drive->get_name());
-      return hardware_interface::return_type::ERROR;
-    }
-
-    // Update the coupled state (one could call this motor readings)
-    const auto& latest_reading = drive->get_last_state();
-    state.position = latest_reading.joint_position;
-    state.velocity = latest_reading.joint_velocity;
-    state.acceleration = latest_reading.joint_acceleration;
-    state.torque = latest_reading.joint_torque;
-  }
-
-  // Now comes the interesting part. We need to take the data from each drive and apply the serial linkage
-  // NOTE: This assumes the joints are declared in the correct order (No clue how I could validate that)
-  kinematics::map_from_coupled_to_serial(state_coupled_kinematics_, state_serial_kinematics_);
-  // Perform the update of the exposed state interface
-  update_state_interfaces(state_interface_mapping_, *this);
-
-  return hardware_interface::return_type::OK;
-}
-hardware_interface::return_type DynaArmHardwareInterface::write([[maybe_unused]] const rclcpp::Time& time,
-                                                                [[maybe_unused]] const rclcpp::Duration& period)
-{
-  // Get commands from the ros2control exposed command interfaces
-  update_command_interfaces(command_interface_mapping_, *this);
-  // Translated commands to coupled kinematics
-  kinematics::map_from_serial_to_coupled(commands_serial_kinematics_, commands_coupled_kinematics_);
-
-  // TODO(firesurfer) port fancy self collision avoidance logic
-
-  // Stage all commands
-  for (std::size_t i = 0; i < drives_.size(); i++) {
-    auto& drive = drives_[i];
-    auto& cmd = commands_coupled_kinematics_[i];
-
-    auto command = drive->get_last_command();
-    command.joint_position = cmd.position;
-    command.joint_velocity = cmd.velocity;
-    command.joint_acceleration = cmd.acceleration;
-    command.joint_torque = cmd.torque;
-
-    drive->stage_command(command);
-  }
-
-  // Perform actual write action
-  // Note this is still asynchronous to the bus communication
-  for (auto& drive : drives_) {
-    if (drive->write() != hardware_interface::return_type::OK) {
-      RCLCPP_ERROR_STREAM(logger_, "Failed to 'write' to drive: " << drive->get_name());
-      return hardware_interface::return_type::ERROR;
-    }
-  }
-  return hardware_interface::return_type::OK;
-}
-
-hardware_interface::return_type DynaArmHardwareInterface::prepare_command_mode_switch(
-    const std::vector<std::string>& start_interfaces, const std::vector<std::string>& stop_interfaces)
-{
-  return hardware_interface::return_type::OK;
-}
-
-hardware_interface::return_type DynaArmHardwareInterface::perform_command_mode_switch(
-    const std::vector<std::string>& start_interfaces, const std::vector<std::string>& stop_interfaces)
-{
-  const auto next_mode = select_mode(start_interfaces, stop_interfaces, logger_);
-  for (auto& drive : drives_) {
-    drive->configure_drive_mode(next_mode);
-  }
-  return hardware_interface::return_type::OK;
-}
-
-DynaArmHardwareInterface::~DynaArmHardwareInterface()
-{
-  RCLCPP_INFO_STREAM(logger_, "Destructor of DynaArm Hardware Interface called");
-}
+using DynaArmHardwareInterface = DynaArmHardwareInterfaceBase<duatic_ros2control_hardware::DuaDriveInterface>;
 }  // namespace duatic_dynaarm_driver
 
 #include "pluginlib/class_list_macros.hpp"
